Refactor to use common util method, add check_mode support
